+++ cli/cli.py
@@ -2,6 +2,7 @@
 import click
 
 from . import api, auth, gcloud, upload, config, consent, __version__
+from .dbedit.cli import list_, remove_
 
 #### $ cidc ####
 @click.group()
@@ -156,6 +157,8 @@ analyses.add_command(upload_analysis)
 admin_.add_command(grant_all)
 admin_.add_command(test_csms)
 admin_.add_command(load_blobs)
+admin_.add_command(list_)
+admin_.add_command(remove_)
 
 if __name__ == "__main__":
     cidc()